Fixed and cleaned the multi agent code
