+++ unified_planning/model/__init__.py
@@ -58,9 +58,7 @@ from unified_planning.model.timing import (
 from unified_planning.model.types import Type, TypeManager
 from unified_planning.model.variable import Variable, FreeVarsOracle
 import unified_planning.model.metrics
-from unified_planning.model.ma_problem import MultiAgentProblem
-from unified_planning.model.agent import Agent
-from unified_planning.model.ma_environment import MAEnvironment
+import unified_planning.model.multi_agent
 
 __all__ = [
     "Action",
@@ -106,7 +104,4 @@ __all__ = [
     "TypeManager",
     "Variable",
     "FreeVarsOracle",
-    "MultiAgentProblem",
-    "Agent",
-    "MAEnvironment",
 ]
