+++ unified_planning/model/fluent.py
@@ -20,6 +20,7 @@ that defines the types of its parameters.
 
 import unified_planning as up
 from unified_planning.environment import get_env, Environment
+from unified_planning.model.types import domain_size, domain_item
 from typing import List, OrderedDict, Optional, Union
 
 
@@ -233,3 +234,41 @@ class Fluent:
 
     def Iff(self, right):
         return self._env.expression_manager.Iff(self, right)
+
+
+def get_ith_fluent_exp(
+    objects_set: "up.model.mixins.ObjectsSetMixin",
+    fluent: "up.model.fluent.Fluent",
+    domain_sizes: List[int],
+    idx: int,
+) -> "up.model.fnode.FNode":
+    """Returns the ith ground fluent expression."""
+    quot = idx
+    rem = 0
+    actual_parameters = []
+    for i, p in enumerate(fluent.signature):
+        ds = domain_sizes[i]
+        rem = quot % ds
+        quot //= ds
+        v = domain_item(objects_set, p.type, rem)
+        actual_parameters.append(v)
+    return fluent(*actual_parameters)
+
+
+def get_all_fluent_exp(
+    objects_set: "up.model.mixins.ObjectsSetMixin",
+    fluent: "up.model.fluent.Fluent",
+) -> List["up.model.fnode.FNode"]:
+    if fluent.arity == 0:
+        return [fluent.environment.expression_manager.FluentExp(fluent)]
+    else:
+        res = []
+        ground_size = 1
+        domain_sizes = []
+        for p in fluent.signature:
+            ds = domain_size(objects_set, p.type)
+            domain_sizes.append(ds)
+            ground_size *= ds
+        for i in range(ground_size):
+            res.append(get_ith_fluent_exp(objects_set, fluent, domain_sizes, i))
+        return res
