+++ unified_planning/model/multi_agent/agent.py
@@ -12,22 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import unified_planning as up
-from unified_planning.environment import get_env, Environment
-from unified_planning.shortcuts import *
-import unified_planning.model.operators as op
-from unified_planning.exceptions import (
-    UPProblemDefinitionError,
-    UPTypeError,
-    UPValueError,
-    UPExpressionDefinitionError,
-)
-from unified_planning.model.walkers import OperatorsExtractor
-from fractions import Fraction
-from typing import List, Dict, Set, Union, Optional
-from unified_planning.model.expression import ConstantExpression
-from unified_planning.model.abstract_problem import AbstractProblem
 from unified_planning.model.mixins import (
     ActionsSetMixin,
     FluentsSetMixin,
@@ -38,27 +23,25 @@ class Agent(
     FluentsSetMixin,
     ActionsSetMixin,
 ):
-
     """Represents an Agent."""
 
     def __init__(
         self,
         name: str,
-        ma_problem: "unified_planning.model.MultiAgentProblem",
+        ma_problem: "up.model.multi_agent.ma_problem.MultiAgentProblem",
     ):
-        self._name: str = name
-        self._ma_problem = ma_problem
-        self._env = up.environment.get_env(self._ma_problem.env)
         FluentsSetMixin.__init__(
             self,
-            self.env,
-            self._ma_problem._add_user_type_method,
+            ma_problem.env,
+            ma_problem._add_user_type,
             self.has_name,
-            self._ma_problem._initial_defaults,
+            ma_problem._initial_defaults,
         )
         ActionsSetMixin.__init__(
-            self, self.env, self._ma_problem._add_user_type_method, self.has_name
+            self, ma_problem.env, ma_problem._add_user_type, self.has_name
         )
+        self._name: str = name
+        self._ma_problem = ma_problem
 
     def has_name(self, name: str) -> bool:
         """Returns true if the name is in the problem."""
@@ -74,17 +57,13 @@ class Agent(
         return self._name
 
     @property
-    def environment(self) -> "Environment":
+    def env(self) -> "up.Environment":
         """Returns the Agent environment."""
-        return self._env
+        return self._ma_problem.env
 
     def __eq__(self, oth: object) -> bool:
         if isinstance(oth, Agent):
-            return (
-                self._name == oth._name
-                and self._ma_problem == oth._ma_problem
-                and self._env == oth._env
-            )
+            return self._name == oth._name and self._ma_problem == oth._ma_problem
         else:
             return False
